switch all controllers to SOA except the groups controllers , edit in saved announcement methods for web
