+++ app/Http/Controllers/AdvancedUserController.php
@@ -1,49 +1,30 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Http\Requests\AdvancedUser\AdvancedUser1;
-use App\Models\AdvancedUser;
-use App\Models\ServiceManager;
-use App\Models\User;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
+use App\Services\AdvancedUserService;
 
 class AdvancedUserController extends Controller
 {
+    protected $advancedUserService;
+
+    public function __construct(AdvancedUserService $advancedUserService)
+    {
+        $this->advancedUserService = $advancedUserService;
+    }
 
     public function showProfile()
     {
-        $user = Auth::user();
-
-        return response()->json($user);
+        return $this->advancedUserService->showProfile();
     }
 
     public function showAll()
     {
-        $users = User::whereHas('advancedUser')->get();
+        $usersData = $this->advancedUserService->showAll();
 
-        $usersData = [];
-
-        foreach ($users as $user) {
-            if ($user->advancedUser->isAccountCompleted == 0) {
-                $usersData[] = [
-                    'id' => $user->advancedUser->id,
-                    'fullName' => $user->fullName,
-                    'email' => $user->email,
-                    'password' => $user->password,
-                    'isAccountCompleted'=>$user->advancedUser->isAccountCompleted
-                ];
-            }
-            else {
-                $usersData[] = [
-                    'id' => $user->advancedUser->id,
-                    'fullName' => $user->fullName,
-                    'email' => $user->email,
-                    'isAccountCompleted'=>$user->advancedUser->isAccountCompleted
-                ];
-            }
-        }
         if(request()->is('api/*')) {
+
             return response()->json($usersData, 200);
         }
         return view('pages.AdvancedUsersTablePageForServiceManager',compact('usersData'));
@@ -51,34 +32,25 @@ class AdvancedUserController extends Controller
 
     public function createAccount(AdvancedUser1 $request)
     {
-        $user =User::query()->create([
-            'fullName'=>$request['fullName'],
-            'password'=>Str::random(12)
-        ]);
-        AdvancedUser::query()->create([
-            'userID'=>$user['id']
-        ]);
+        $user = $this->advancedUserService->createAccount($request);
+
         if(request()->is('api/*')) {
+
             return response()->json($user);
         }
         return redirect()->back();
-       // return view('');
     }
 
     // complete account = set email in AuthController
 
     public function deleteAllAccounts()
     {
-        if (ServiceManager::where('userID', auth()->id())->where('position', 'provost')->exists()) {
+        $response = $this->advancedUserService->deleteAllAccounts();
 
-            User::whereHas('advancedUser')->delete();
+        if(request()->is('api/*')) {
 
-            if(request()->is('api/*')) {
-                return response()->json(['message' => 'All Accounts have been Deleted Successfully']);
-            }
-            return redirect()->back();
+            return response()->json($response);
         }
-        return response()->json(['message' => 'you dont have the permission to delete all records in this table']);
+        return redirect()->back();
     }
-
 }
