+++ app/Http/Controllers/AnnouncementController.php
@@ -7,19 +7,24 @@ use App\Http\Requests\Announcement\Announcement2;
 use App\Http\Requests\Service\Service4;
 use App\Models\Announcement;
 use App\Models\Service;
-use App\Models\User;
-use App\Notifications\AnnouncementNotification;
+use App\Services\AnnouncementService;
 use Illuminate\Http\Response;
-use Illuminate\Support\Facades\Notification;
 
 class AnnouncementController extends Controller
 {
+    protected $announcementService;
+
+    public function __construct(AnnouncementService $announcementService)
+    {
+        $this->announcementService = $announcementService;
+    }
 
     public function showAll()
     {
-        $allRecords = Announcement::with('service', 'file')->get();
+        $allRecords = $this->announcementService->showAll();
 
         if (request()->is('api/*')) {
+
             return response()->json($allRecords, Response::HTTP_OK);
         }
         return view('pages.AdvertismentPageFromServiceForServiceManage',compact('allRecords'));
@@ -27,115 +32,67 @@ class AnnouncementController extends Controller
 
     public function showAllFromService(Service $service)
     {
-        $allRecords = Announcement::where('serviceID', $service['id'])->with('service', 'file')->get();
+        $allRecords = $this->announcementService->showAllFromService($service);
 
         if (request()->is('api/*')) {
+
             return response()->json($allRecords, Response::HTTP_OK);
         }
-        return view('');
+        return view('',compact('allRecords'));
     }
 
     public function showMy()
     {
-        $allRecords = Announcement::where('userID', auth()->id())->with('service', 'file')->get();
+        $allRecords = $this->announcementService->showMy();
 
         if (request()->is('api/*')) {
+
             return response()->json($allRecords, Response::HTTP_OK);
         }
-        return view('');
+        return view('',compact('allRecords'));
     }
 
     public function add(Announcement1 $request)
     {
-        $data = $request->validated();
-
-        $data['userID'] = auth()->id();
-
-        $announcement = Announcement::create($data);
-
-        if($request['file']) {
-            $announcement['fileStored'] = $this->addFileInAnnouncement($request,  $announcement);
-        }
-
-        //////////////////// Send notification to all users except the one with id=1
-
-        $users = User::where('id', '!=', 1)->get();
-
-        Notification::send($users, new AnnouncementNotification($announcement));
-
-        $latestNotification = auth()->user()->notifications()->latest()->first();
+        $latestNotification = $this->announcementService->add($request);
 
         if (request()->is('api/*')) {
-            return response()->json([$announcement, $latestNotification->data], Response::HTTP_OK);
+
+            return response()->json($latestNotification->data, Response::HTTP_OK);
         }
-        return redirect()->back()->with([
-            'notificationData' => $latestNotification->data
-        ]);
+        return redirect()->back()->with(['notificationData' => $latestNotification->data]);
     }
 
     public function addFromService(Announcement2 $request, Service $service)
     {
-        $data = $request->validated();
-
-        $data['serviceID'] = $service['id'];
-
-        $data['userID'] = auth()->id();
-
-        $announcement = Announcement::create($data);
-
-        if($request['file']) {
-            $announcement['fileStored'] = $this->addFileFromServiceInAnnouncement($request,  $announcement);
-        }
-
-        //////////////////// Send notification to all users except the one with id=1
-
-        $users = User::where('id', '!=', 1)->get();
-
-        Notification::send($users, new AnnouncementNotification($announcement));
-
-        $latestNotification = auth()->user()->notifications()->latest()->first();
+        $latestNotification = $this->announcementService->addFromService($request, $service);
 
         if (request()->is('api/*')) {
-            return response()->json([$announcement, $latestNotification->data], Response::HTTP_OK);
+
+            return response()->json($latestNotification->data, Response::HTTP_OK);
         }
-        return redirect()->back()->with([
-            'notificationData' => $latestNotification->data
-        ]);
+        return redirect()->back()->with(['notificationData' => $latestNotification->data]);
     }
 
     public function update(Announcement1 $request, Announcement $announcement)
     {
-        $recordUpdated = $announcement->update($request->validated());
+        $recordUpdated = $this->announcementService->update($request, $announcement);
 
         if (request()->is('api/*')) {
+
             return response()->json($recordUpdated, Response::HTTP_OK);
         }
-        return view('');
+        return redirect()->back();
     }
 
     public function filterByType(Service4 $request)
     {
-        $filterType = $request['filterType'];
-        $filterName = $request['filterName'];
-
-        $query = Announcement::query();
-
-        if ($filterType == 'serviceYear' || $filterType == 'serviceSpecializationName') {
-            $query->whereHas('service.serviceYearAndSpecialization', function ($subQuery) use ($filterType, $filterName) {
-                $subQuery->where($filterType, $filterName);
-            });
-        }
-        elseif ($filterType == 'serviceName' || $filterType == 'serviceType') {
-            $query->whereHas('service', function ($subQuery) use ($filterType, $filterName) {
-                $subQuery->where($filterType, $filterName);
-            });
-        }
-        $filteredServices = $query->with('service', 'file')->get();
+        $filteredServices = $this->announcementService->filterByType($request);
 
         if (request()->is('api/*')) {
+
             return response()->json($filteredServices, Response::HTTP_OK);
         }
-        return view('');
+        return view('',compact('filteredServices'));
     }
-
 }
