+++ app/Http/Controllers/Controller.php
@@ -5,192 +5,98 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Announcement\Announcement1;
 use App\Http\Requests\Announcement\Announcement2;
 use App\Http\Requests\File\File1;
-use App\Http\Requests\User\User3;
-use App\Mail\VerifyEmail;
-use App\Models\FakeReservation;
-use App\Models\Session;
 use App\Models\User;
-use Carbon\Carbon;
+use App\Services\ControllerService;
 use App\Models\Announcement;
-use App\Models\File;
-use App\Models\InterestedService;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Validation\ValidatesRequests;
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 use Illuminate\Routing\Controller as BaseController;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Mail;
-use PharIo\Manifest\Email;
 use App\Http\Requests\User\User4;
-use Maatwebsite\Excel\Facades\Excel;
 
 class Controller extends BaseController
 {
     use AuthorizesRequests, ValidatesRequests;
 
+    protected $controllerService;
+
+    public function __construct(ControllerService $controllerService)
+    {
+        $this->controllerService = $controllerService;
+    }
+
     public function sendEmail($email)
     {
-        $verify = DB::table('password_reset_tokens')->where([['email',$email]]);
-
-        if ($verify->exists()) {
-            $verify->delete();
-        }
-
-        $pin = rand(100000, 999999);
-
-        DB::table('password_reset_tokens')->insert(['email' => $email, 'token' => $pin ]);
-
-        Mail::to($email)->send(new VerifyEmail($pin));
+        $this->controllerService->sendEmail($email);
     }
 
     public function createToken(User $user)
     {
-        $tokenResult = $user->createToken('Personal Access Token');
-
-        $data["user"]= $user;
-
-        $data["token_type"]='Bearer';
-
-        $data["access_token"]=$tokenResult->accessToken;
-
-        return $data;
+        return $this->controllerService->createToken($user);
     }
 
     public function checkToken(User4 $request)
     {
-        $select = DB::table('password_reset_tokens')->where('token', $request['token']);
-
-        if ($select->get()->isEmpty()) {
-            return false;
-        }
-        return $select->value('email');
+        return $this->controllerService->checkToken($request);
     }
 
-    protected function importUsersFile(File1 $request, $importClass)
+    public function importUsersFile(File1 $request, $importClass)
     {
-        if ($request['file']->isValid()) {
-            Excel::import(new $importClass(), $request['file']);
+         $response = $this->controllerService->importUsersFile($request, $importClass);
 
-//            if(request()->is('api/*')) {
-                return response()->json(['message' => 'File imported successfully'], 200);
-//            }
-        }
-        else {
-//            if(request()->is('api/*')) {
-                return response()->json(['message' => 'Failed to upload file'], 500);
-//            }
-        }
+         if(request()->is('api/*')) {
+
+             return response()->json($response);
+         }
+         return redirect()->back();
     }
 
-    function checkIsInterested($serviceId, $userId)
+    public function checkIsInterested($serviceId, $userId)
     {
-        $exists = InterestedService::where('serviceID', $serviceId)
-            ->where('userID', $userId)
-            ->exists();
-
-        return $exists ? 1 : 0;
+        return $this->controllerService->checkIsInterested($serviceId, $userId);
     }
 
-    protected function getServiceData($allRecords)
+    public function getServiceData($allRecords)
     {
-        return $allRecords->map(function ($record) {
-            $advancedUserRoles = $record->assignedService->map(function ($assignedService) {
-                return [
-                    'fullName' => $assignedService->user->fullName,
-                    'roles' => $assignedService->assignedRole->pluck('role.roleName')
-                ];
-            });
+        return $this->controllerService->getServiceData($allRecords);
+    }
 
-            $interestedServices = $record->interestedService;
+    public function checkIsSaved($announcementId, $userId)
+    {
+        return $this->controllerService->checkIsSaved($announcementId, $userId);
+    }
 
-            $isInterested = $interestedServices->contains(function ($interestedService) use ($record) {
-                return $this->checkIsInterested($record['id'], $interestedService->userID);
-            });
-
-            if($isInterested){
-                $interestedService= $interestedServices->firstWhere('userID', auth()->id());
-            }else{
-                $interestedService=null;
-            }
-
-
-            return [
-                'id' => $record['id'],
-                'serviceManagerName' => $record->serviceManager->user->fullName,
-                'parentServiceID' =>$record['parentServiceID'],
-                'parentServiceName' => $record->parentService?->serviceName,
-                'serviceYearName' => $record->serviceYearAndSpecialization->serviceYear,
-                'serviceSpecializationName' => $record->serviceYearAndSpecialization->serviceSpecializationName,
-                'serviceName' => $record->serviceName,
-                'serviceDescription' => $record->serviceDescription,
-                'serviceType' => $record->serviceType,
-                'minimumNumberOfGroupMembers' => $record->minimumNumberOfGroupMembers,
-                'maximumNumberOfGroupMembers' => $record->maximumNumberOfGroupMembers,
-                'statusName' => $record->status == 1 ? 'Effective' : 'Not Effective',
-                'advancedUsersWithRoles' => $advancedUserRoles,
-                'isInterested' => $isInterested ? 1 : 0,
-                'interestedService' => $interestedService
-            ];
-        });
+    public function getAnnouncementData($allRecords)
+    {
+        return $this->controllerService->getAnnouncementData($allRecords);
     }
 
     public function createFakeReservations($session): array
     {
-        $privateSession = $session->privateSession;
-        $sessionStartTime = Carbon::parse($session->sessionStartTime);
-        $sessionEndTime = Carbon::parse($session->sessionEndTime);
-        $durationForEachReservation = Carbon::parse($privateSession->durationForEachReservation);
-
-        $totalSessionDurationMinutes = $sessionEndTime->diffInMinutes($sessionStartTime);
-        $reservationDurationMinutes = $durationForEachReservation->hour * 60 + $durationForEachReservation->minute;
-        $numberOfReservations = intval($totalSessionDurationMinutes / $reservationDurationMinutes);
-
-        $reservations = [];
-        for ($i = 0; $i < $numberOfReservations; $i++) {
-            $reservationStartTime = $sessionStartTime->copy()->addMinutes($i * $reservationDurationMinutes);
-            $reservationEndTime = $reservationStartTime->copy()->addMinutes($reservationDurationMinutes);
-
-            $reservationStartTimeFormatted = $reservationStartTime->format('H:i');
-            $reservationEndTimeFormatted = $reservationEndTime->format('H:i');
-
-            $reservations[] = FakeReservation::create([
-                'privateSessionID' => $privateSession->id,
-                'reservationStartTime' => $reservationStartTimeFormatted,
-                'reservationEndTime' => $reservationEndTimeFormatted
-            ]);
-        }
-
-        return $reservations;
+        return $this->controllerService->createFakeReservations($session);
     }
 
     private function storeFile($request, $announcement)
     {
-        $newFile = $request->file('file');
-
-        $data['announcementID'] = $announcement->id;
-
-        $data['fileName'] = $newFile->getClientOriginalName();
-
-        $data['filePath'] = $newFile->storeAs('uploads', $data['fileName'], 'public');
-
-        return File::create($data);
+        return $this->controllerService->storeFile($request, $announcement);
     }
 
     public function addFileInAnnouncement(Announcement1 $request, Announcement $announcement)
     {
-        $recordStored = $this->storeFile($request, $announcement);
-
-        return response()->json($recordStored, Response::HTTP_OK);
+        return $this->controllerService->addFileInAnnouncement($request, $announcement);
     }
 
     public function addFileFromServiceInAnnouncement(Announcement2 $request, Announcement $announcement)
     {
-        $recordStored = $this->storeFile($request, $announcement);
-
-        return response()->json($recordStored, Response::HTTP_OK);
+        return $this->controllerService->addFileFromServiceInAnnouncement($request, $announcement);
     }
 
-}
+    public function showMyReservationsByType($serviceType)
+    {
+        return $this->controllerService->showMyReservationsByType($serviceType);
+    }
 
+    public function getSwapRequestsDetails($swapRequests)
+    {
+        return $this->controllerService->getSwapRequestsDetails($swapRequests);
+    }
+}
