+++ app/Http/Controllers/InterestedServiceController.php
@@ -4,42 +4,31 @@ namespace App\Http\Controllers;
 
 use App\Models\InterestedService;
 use App\Models\Service;
+use App\Services\InterestedServiceService;
 use Illuminate\Http\Response;
 
 class InterestedServiceController extends Controller
 {
+    protected $interestedServiceService;
+
+    public function __construct(InterestedServiceService $interestedServiceService)
+    {
+        $this->interestedServiceService = $interestedServiceService;
+    }
 
     public function showAll()
     {
-        $allRecords = InterestedService::whereHas('service', function ($query) {
-            $query->whereNull('parentServiceID')
-                ->where('status', 1);
-        })
-            ->where('userID', auth()->id())
-            ->with('service')
-            ->get();
+        $allRecords = $this->interestedServiceService->showAll();
 
-        foreach ($allRecords as $record) {
-            $record->service['children'] = Service::where('parentServiceID', $record->service['id'])
-                ->where('status', 1)
-                ->get();
-        }
         return response()->json($allRecords, Response::HTTP_OK);
     }
 
     public function showAllParent()
     {
-        $allRecords = Service::with(['interestedService', 'serviceManager.user', 'parentService', 'serviceYearAndSpecialization', 'assignedService.user', 'assignedService.assignedRole.role'])
-            ->whereHas('interestedService', function ($query) {
-                $query->where('userID', auth()->id());
-            })
-            ->whereNull('parentServiceID')
-            ->orderByDesc('status')
-            ->get();
-
-        $allRecords = $this->getServiceData($allRecords);
+        $allRecords = $this->interestedServiceService->showAllParent();
 
         if (request()->is('api/*')) {
+
             return response()->json($allRecords, Response::HTTP_OK);
         }
         return view('pages.FavoritePublicServicesPageForServiceManager',[
@@ -49,14 +38,10 @@ class InterestedServiceController extends Controller
 
     public function showChild(Service $service)
     {
-        $allRecords = Service::with(['interestedService', 'serviceManager.user', 'parentService', 'serviceYearAndSpecialization', 'assignedService.user', 'assignedService.assignedRole.role'])
-            ->where('parentServiceID', $service['id'])
-            ->orderByDesc('status')
-            ->get();
-
-        $allRecords = $this->getServiceData($allRecords);
+        $allRecords = $this->interestedServiceService->showChild($service);
 
         if (request()->is('api/*')) {
+
             return response()->json($allRecords, Response::HTTP_OK);
         }
         return view('pages.FavoritePrivateServicesPageForServiceManager', [
@@ -67,25 +52,23 @@ class InterestedServiceController extends Controller
 
     public function interestIn(Service $service)
     {
-        $data['userID'] = auth()->id();
-        $data['serviceID'] = $service['id'];
-
-        $recordStored = InterestedService::create($data);
+        $recordStored = $this->interestedServiceService->interestIn($service);
 
         if (request()->is('api/*')) {
+
             return response()->json($recordStored, Response::HTTP_OK);
         }
-        return view('');
+        return redirect()->back();
     }
 
     public function unInterestIn(InterestedService $interestedService)
     {
-        $interestedService->delete();
+        $response = $this->interestedServiceService->unInterestIn($interestedService);
 
         if (request()->is('api/*')) {
-            return response()->json(['message' => 'this service uninterested successfully']);
-        }
-        return view('');
-    }
 
+            return response()->json($response);
+        }
+        return redirect()->back();
+    }
 }
