+++ app/Http/Controllers/SavedAnnouncementController.php
@@ -4,43 +4,48 @@ namespace App\Http\Controllers;
 
 use App\Models\Announcement;
 use App\Models\SavedAnnouncement;
+use App\Services\SavedAnnouncementService;
 use Illuminate\Http\Response;
 
 class SavedAnnouncementController extends Controller
 {
+    protected $savedAnnouncementService;
+
+    public function __construct(SavedAnnouncementService $savedAnnouncementService)
+    {
+        $this->savedAnnouncementService = $savedAnnouncementService;
+    }
 
     public function showAll()
     {
-        $allRecords = SavedAnnouncement::where('userID', auth()->id())
-            ->with('announcement', 'announcement.service', 'announcement.file')->get();
+        $allRecords = $this->savedAnnouncementService->showAll();
 
         if (request()->is('api/*')) {
+
             return response()->json($allRecords, Response::HTTP_OK);
         }
-        return view('');
+        return view('',compact('allRecords'));
     }
 
     public function save(Announcement $announcement)
     {
-        $data['userID'] = auth()->id();
-        $data['announcementID'] = $announcement['id'];
-
-        $recordStored = SavedAnnouncement::create($data);
+        $recordStored = $this->savedAnnouncementService->save($announcement);
 
         if (request()->is('api/*')) {
+
             return response()->json($recordStored, Response::HTTP_OK);
         }
-        return view('');
+        return Redirect()->back();
     }
 
     public function unSave(SavedAnnouncement $savedAnnouncement)
     {
-        $savedAnnouncement->delete();
+        $response = $this->savedAnnouncementService->unSave($savedAnnouncement);
 
         if (request()->is('api/*')) {
-            return response()->json(['message' => 'this announcement unsaved successfully']);
-        }
-        return view('');
-    }
 
+            return response()->json($response);
+        }
+        return Redirect()->back();
+    }
 }
