--- app/Http/Controllers/SwapRequestController.php
@@ -1,39 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\SwapRequest;
-use Illuminate\Http\Request;
-
-class SwapRequestController extends Controller
-{
-
-    public function index()
-    {
-        //
-    }
-
-
-    public function store(Request $request)
-    {
-        //
-    }
-
-
-    public function show(SwapRequest $swapRequest)
-    {
-        //
-    }
-
-
-    public function update(Request $request, SwapRequest $swapRequest)
-    {
-        //
-    }
-
-
-    public function destroy(SwapRequest $swapRequest)
-    {
-        //
-    }
-}
