+++ app/Notifications/AnnouncementNotification.php
@@ -38,7 +38,6 @@ class AnnouncementNotification extends Notification
     public function toDatabase($notifiable)
     {
         return [
-            'announcementID' => $this->announcement->id,
             'serviceName' => $this->announcement->service->serviceName ?? null,
             'title' => $this->announcement->title
         ];
