+++ app/Services/FakeReservationService.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Services;
+
+use App\Models\PrivateSession;
+
+class FakeReservationService
+{
+    public function showALl(PrivateSession $privateSession)
+    {
+        return $privateSession->fakeReservation;
+    }
+}
