+++ app/Services/NotificationService.php
@@ -2,7 +2,6 @@
 
 namespace App\Services;
 
-use App\Models\Announcement;
 use Illuminate\Notifications\DatabaseNotification;
 use Illuminate\Support\Facades\Auth;
 
@@ -13,19 +12,12 @@ class NotificationService
         return Auth::user()->notifications;
     }
 
-    public function markNotificationAsRead($id)
+    public function markAsRead($id)
     {
         $notification = DatabaseNotification::findOrFail($id);
 
         $notification->markAsRead();
 
         return ['message' => 'Notification marked as read'];
-    }
-
-    public function goToAnnouncementFromNotification($id)
-    {
-        $notification = DatabaseNotification::findOrFail($id);
-
-        return Announcement::where('id', $notification->data['announcementID'])->first();;
     }
 }
