+++ composer.json
@@ -8,6 +8,7 @@
         "php": "^8.1",
         "endroid/qr-code": "^5.0",
         "guzzlehttp/guzzle": "^7.8",
+        "kreait/firebase-php": "^7.13",
         "kriswallsmith/buzz": "^0.15.0",
         "laravel/framework": "^10.0",
         "laravel/passport": "^12.2",
